Refactoring utilities to CommonController.cs
Utility functions for handling PMU/PDC devices have been extracted from DeviceController.cs and moved to CommonController.cs, centralizing connection logic, parsing, retry, and measurement handling. DeviceController.cs now utilizes these shared methods, reducing duplication and improving maintainability. The project has been updated to compile the new CommonController.cs file.
